# M2-Challenge-Lim-Blaze/src/main/java/com/company/M2ChallengeLimBlaze/models/MathSolution.java
package com.company.M2ChallengeLimBlaze.models;

import com.fasterxml.jackson.annotation.JsonSubTypes;

import javax.validation.constraints.Digits;
import javax.validation.constraints.NotEmpty;
import java.util.Objects;

public class MathSolution {

    @NotEmpty(message = "You must supply a value for operand1")
    private Integer operand1;

    @NotEmpty(message = "You must supply a value for operand2")
    private Integer operand2;
    private String operation;
    private Integer answer;

    public MathSolution(Integer operand1, Integer operand2) {
        this.operand1 = operand1;
        this.operand2 = operand2;
    }

    public MathSolution() {
    }

    public void setOperand1(Integer operand1) {
        this.operand1 = operand1;
    }

    public void setOperand2(Integer operand2) {
        this.operand2 = operand2;
    }

    public String getOperation() {
        return operation;
    }

    public void setOperation(String operation) {
        this.operation = operation;
    }

    public Integer getAnswer() {
        return answer;
    }

    public void setAnswer(Integer answer) {
        this.answer = answer;
    }

    public Integer getOperand1() {
        return operand1;
    }

    public Integer getOperand2() {
        return operand2;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        MathSolution that = (MathSolution) o;
        return Objects.equals(operand1, that.operand1) && Objects.equals(operand2, that.operand2) && Objects.equals(operation, that.operation) && Objects.equals(answer, that.answer);
    }

    @Override
    public int hashCode() {
        return Objects.hash(operand1, operand2, operation, answer);
    }
}